+++ internal/repository/models/filters.go
@@ -1,0 +1,46 @@
+package models
+
+import (
+	"pasteAPI/pkg/validator"
+	"strings"
+)
+
+type Filters struct {
+	Page         uint32
+	PageSize     uint32
+	Sort         string
+	SortSafelist []string
+}
+
+func ValidateFilters(v *validator.Validator, f Filters) {
+	// Check that the page and page_size parameters contain sensible values.
+	v.Check(f.Page > 0, "page", "must be greater than zero")
+	v.Check(f.Page <= 10_000_000, "page", "must be a maximum of 10 million")
+	v.Check(f.PageSize > 0, "page_size", "must be greater than zero")
+	v.Check(f.PageSize <= 100, "page_size", "must be a maximum of 100")
+	// Check that the sort parameter matches a value in the safelist.
+	v.Check(validator.In(f.Sort, f.SortSafelist...), "sort", "invalid sort value")
+}
+
+func (f Filters) SortColumn() string {
+	for _, safeValue := range f.SortSafelist {
+		if f.Sort == safeValue {
+			return strings.TrimPrefix(f.Sort, "-")
+		}
+	}
+	panic("unsafe sort parameter: " + f.Sort)
+}
+
+func (f Filters) SortDirection() string {
+	if strings.HasPrefix(f.Sort, "-") {
+		return "DESC"
+	}
+	return "ASC"
+}
+
+func (f Filters) Limit() uint32 {
+	return f.PageSize
+}
+func (f Filters) Offset() uint32 {
+	return (f.Page - 1) * f.PageSize
+}
