+++ internal/repository/models/password.go
@@ -1,0 +1,34 @@
+package models
+
+import (
+	"errors"
+	"golang.org/x/crypto/bcrypt"
+)
+
+type password struct {
+	Plaintext *string
+	Hash      []byte
+}
+
+func (p *password) Set(plaintextPassword string) error {
+	hash, err := bcrypt.GenerateFromPassword([]byte(plaintextPassword), 12)
+	if err != nil {
+		return err
+	}
+	p.Plaintext = &plaintextPassword
+	p.Hash = hash
+	return nil
+}
+
+func (p *password) Matches(plaintextPassword string) (bool, error) {
+	err := bcrypt.CompareHashAndPassword(p.Hash, []byte(plaintextPassword))
+	if err != nil {
+		switch {
+		case errors.Is(err, bcrypt.ErrMismatchedHashAndPassword):
+			return false, nil
+		default:
+			return false, err
+		}
+	}
+	return true, nil
+}
