+++ internal/repository/models/paste.go
@@ -1,6 +1,9 @@
 package models
 
-import "time"
+import (
+	"pasteAPI/pkg/validator"
+	"time"
+)
 
 type Paste struct {
 	Id        uint16    `json:"id"`
@@ -11,4 +14,14 @@ type Paste struct {
 	ExpiresAt time.Time `json:"expires_at"`
 	Minutes   int32     `json:"-"`
 	Version   uint32    `json:"version"`
+}
+
+func ValidatePaste(v *validator.Validator, p *Paste) {
+	v.Check(p.Title != "", "title", "must be provided")
+	v.Check(len(p.Title) <= 255, "title", "must not be more than 500 bytes long")
+
+	v.Check(CategoriesList.IsValidCategory(p.Category), "category", "no such category")
+
+	v.Check(p.Text != "", "text", "must be provided")
+	v.Check(len(p.Title) <= 500, "title", "must not be more than 500 bytes long")
 }
